Fix deserialization of burst_colours on MessageReaction
